requirements.txt
# python 3.11

PyYAML==6.0.1
google-generativeai==0.3.2
chromadb==0.4.22